CHANGE: migrate online_boutique example to cpp
