+++ src/include/http.h
@@ -261,6 +261,11 @@ typedef struct _adresponse
     Ad Ads[10];
 } AdResponse;
 
+struct c{
+    AdResponse a;
+    AdRequest b;
+
+};
 struct http_transaction
 {
     uint32_t tenant_id;
@@ -285,6 +290,7 @@ struct http_transaction
     uint32_t length_response;
 
     char rpc_handler[64];
+    // enum evnet_type;
     char caller_nf[64];
     char request[HTTP_MSG_LENGTH_MAX];
     char response[HTTP_MSG_LENGTH_MAX];
