+++ src/online_boutique/paymentservice.cpp
@@ -26,183 +26,158 @@
 #include <sys/epoll.h>
 #include <time.h>
 #include <unistd.h>
+#include <uuid/uuid.h>
 
 #include <rte_branch_prediction.h>
 #include <rte_eal.h>
 #include <rte_errno.h>
 #include <rte_memzone.h>
 
-#include "c_lib.h"
 #include "http.h"
 #include "io.h"
 #include "spright.h"
-#include "utility.h"
 
 static int pipefd_rx[UINT8_MAX][2];
 static int pipefd_tx[UINT8_MAX][2];
 
-char *product_ids[] = {"OLJCESPC7Z", "66VCHSJNUP", "1YMWWN1N4O", "L9ECAV7KIM", "2ZYFJ3GM2N",
-                       "0PUK6V6EV0", "LS4PSXUNUM", "9SIQT8TOJO", "6E92ZMYYFZ"};
-
-Product products[9] = {
-    {.Id = "OLJCESPC7Z",
-     .Name = "Sunglasses",
-     .Description = "Add a modern touch to your outfits with these sleek aviator sunglasses.",
-     .Picture = "/static/img/products/sunglasses.jpg",
-     .PriceUsd = {.CurrencyCode = "USD", .Units = 19, .Nanos = 990000000},
-     .num_categories = 1,
-     .Categories = {"accessories"}},
-    {.Id = "66VCHSJNUP",
-     .Name = "Tank Top",
-     .Description = "Perfectly cropped cotton tank, with a scooped neckline.",
-     .Picture = "/static/img/products/tank-top.jpg",
-     .PriceUsd = {.CurrencyCode = "USD", .Units = 18, .Nanos = 990000000},
-     .num_categories = 2,
-     .Categories = {"clothing", "tops"}},
-    {.Id = "1YMWWN1N4O",
-     .Name = "Watch",
-     .Description = "This gold-tone stainless steel watch will work with most of your outfits.",
-     .Picture = "/static/img/products/watch.jpg",
-     .PriceUsd = {.CurrencyCode = "USD", .Units = 109, .Nanos = 990000000},
-     .num_categories = 1,
-     .Categories = {"accessories"}},
-    {.Id = "L9ECAV7KIM",
-     .Name = "Loafers",
-     .Description = "A neat addition to your summer wardrobe.",
-     .Picture = "/static/img/products/loafers.jpg",
-     .PriceUsd = {.CurrencyCode = "USD", .Units = 89, .Nanos = 990000000},
-     .num_categories = 1,
-     .Categories = {"footwear"}},
-    {.Id = "2ZYFJ3GM2N",
-     .Name = "Hairdryer",
-     .Description = "This lightweight hairdryer has 3 heat and speed settings. It's perfect for travel.",
-     .Picture = "/static/img/products/hairdryer.jpg",
-     .PriceUsd = {.CurrencyCode = "USD", .Units = 24, .Nanos = 990000000},
-     .num_categories = 2,
-     .Categories = {"hair", "beauty"}},
-    {.Id = "0PUK6V6EV0",
-     .Name = "Candle Holder",
-     .Description = "This small but intricate candle holder is an excellent gift.",
-     .Picture = "/static/img/products/candle-holder.jpg",
-     .PriceUsd = {.CurrencyCode = "USD", .Units = 18, .Nanos = 990000000},
-     .num_categories = 2,
-     .Categories = {"decor", "home"}},
-    {.Id = "LS4PSXUNUM",
-     .Name = "Salt & Pepper Shakers",
-     .Description = "Add some flavor to your kitchen.",
-     .Picture = "/static/img/products/salt-and-pepper-shakers.jpg",
-     .PriceUsd = {.CurrencyCode = "USD", .Units = 18, .Nanos = 490000000},
-     .num_categories = 1,
-     .Categories = {"kitchen"}},
-    {.Id = "9SIQT8TOJO",
-     .Name = "Bamboo Glass Jar",
-     .Description = "This bamboo glass jar can hold 57 oz (1.7 l) and is perfect for any kitchen.",
-     .Picture = "/static/img/products/bamboo-glass-jar.jpg",
-     .PriceUsd = {.CurrencyCode = "USD", .Units = 5, .Nanos = 490000000},
-     .num_categories = 1,
-     .Categories = {"kitchen"}},
-    {.Id = "6E92ZMYYFZ",
-     .Name = "Mug",
-     .Description = "A simple mug with a mustard interior.",
-     .Picture = "/static/img/products/mug.jpg",
-     .PriceUsd = {.CurrencyCode = "USD", .Units = 8, .Nanos = 990000000},
-     .num_categories = 1,
-     .Categories = {"kitchen"}}};
-
-static int compare_e(void *left, void *right)
+static int get_digits(int64_t num)
 {
-    return strcmp((const char *)left, (const char *)right);
+    // returns the number of digits
+    return (int)floor(log10(num));
 }
 
-struct clib_map *productcatalog_map;
-
-static void parseCatalog(struct clib_map *map)
+static int get_digit_sum(int n)
 {
-    int size = sizeof(product_ids) / sizeof(product_ids[0]);
-    int i = 0;
-    for (i = 0; i < size; i++)
+    return (int)(n / 10) + (n % 10);
+}
+
+static char *creditcard_validator(int64_t credit_card)
+{
+
+    int digits = get_digits(credit_card);
+    int sum = 0;
+    int first_digits = 0;
+    char *card_type;
+    int i;
+    digits++;
+
+    for (i = 0; i < digits; i++)
     {
-        char *key = clib_strdup(product_ids[i]);
-        int key_length = (int)strlen(key) + 1;
-        Product value = products[i];
-        log_info("Inserting [%s -> %s]", key, value.Name);
-        insert_c_map(map, key, key_length, &value, sizeof(Product));
-        free(key);
-    }
-}
-
-static void ListProducts(struct http_transaction *txn)
-{
-    ListProductsResponse *out = &txn->list_products_response;
-
-    int size = sizeof(out->Products) / sizeof(out->Products[0]);
-    int i = 0;
-    out->num_products = 0;
-    for (i = 0; i < size; i++)
-    {
-        out->Products[i] = products[i];
-        out->num_products++;
-    }
-    return;
-}
-
-static void MockGetProductRequest(struct http_transaction *txn)
-{
-    GetProductRequest *req = &txn->get_product_request;
-    strcpy(req->Id, "2ZYFJ3GM2N");
-}
-
-static void GetProduct(struct http_transaction *txn)
-{
-    GetProductRequest *req = &txn->get_product_request;
-
-    Product *found = &txn->get_product_response;
-    int num_products = 0;
-
-    int size = sizeof(product_ids) / sizeof(product_ids[0]);
-    int i = 0;
-    for (i = 0; i < size; i++)
-    {
-        if (strcmp(req->Id, product_ids[i]) == 0)
+        if (i & 1)
         {
-            log_info("Get Product: %s", product_ids[i]);
-            num_products++;
-            *found = products[i];
-            break;
+            sum += get_digit_sum(2 * (credit_card % 10));
+        }
+        else
+        {
+            sum += credit_card % 10;
+        }
+
+        if (i == digits - 2)
+        {
+            first_digits = credit_card % 10;
+        }
+        else if (i == digits - 1)
+        {
+            first_digits = first_digits + (credit_card % 10) * 10;
+        }
+
+        credit_card /= 10;
+    }
+
+    if (!(sum % 10))
+    {
+        if (digits == 15 && (first_digits == 34 || first_digits == 37))
+        {
+            card_type = "amex";
+        }
+        else if (digits == 16 &&
+                 ((first_digits >= 50 && first_digits <= 55) || (first_digits >= 22 && first_digits <= 27)))
+        {
+            card_type = "mastercard";
+        }
+        else if ((digits >= 13 && digits <= 16) && (first_digits / 10 == 4))
+        {
+            card_type = "visa";
+        }
+        else
+        {
+            card_type = "invalid";
         }
     }
-
-    if (num_products == 0)
+    else
     {
-        log_info("no product with ID %s", req->Id);
+        card_type = "invalid";
     }
+
+    return card_type;
+}
+
+static void Charge(struct http_transaction *txn)
+{
+    log_info("[Charge] received request");
+    ChargeRequest *in = &txn->charge_request;
+
+    Money *amount = &in->Amount;
+    char *cardNumber = in->CreditCard.CreditCardNumber;
+
+    char *cardType;
+    bool valid = false;
+    cardType = creditcard_validator(strtoll(cardNumber, NULL, 10));
+    if (strcmp(cardType, "invalid"))
+    {
+        valid = true;
+    }
+
+    if (!valid)
+    { // throw InvalidCreditCard
+        log_info("Credit card info is invalid");
+        return;
+    }
+
+    // Only VISA and mastercard is accepted,
+    // other card types (AMEX, dinersclub) will
+    // throw UnacceptedCreditCard error.
+    if ((strcmp(cardType, "visa") != 0) && (strcmp(cardType, "mastercard") != 0))
+    {
+        log_info("Sorry, we cannot process %s credit cards. Only VISA or MasterCard is accepted.", cardType);
+        return;
+    }
+
+    // Also validate expiration is > today.
+    int32_t currentMonth = 5;
+    int32_t currentYear = 2022;
+    int32_t year = in->CreditCard.CreditCardExpirationYear;
+    int32_t month = in->CreditCard.CreditCardExpirationMonth;
+    if ((currentYear * 12 + currentMonth) > (year * 12 + month))
+    { // throw ExpiredCreditCard
+        log_info("Your credit card (ending %s) expired on %d/%d", cardNumber, month, year);
+        return;
+    }
+
+    log_info("Transaction processed: %s ending %s Amount: %s%ld.%d", cardType, cardNumber, amount->CurrencyCode,
+             amount->Units, amount->Nanos);
+    uuid_t binuuid;
+    uuid_generate_random(binuuid);
+    uuid_unparse(binuuid, txn->charge_response.TransactionId);
+
     return;
 }
 
-static void MockSearchProductsRequest(struct http_transaction *txn)
+static void MockChargeRequest(struct http_transaction *txn)
 {
-    SearchProductsRequest *req = &txn->search_products_request;
-    strcpy(req->Query, "outfits");
+    strcpy(txn->charge_request.CreditCard.CreditCardNumber, "4432801561520454");
+    txn->charge_request.CreditCard.CreditCardCvv = 672;
+    txn->charge_request.CreditCard.CreditCardExpirationYear = 2039;
+    txn->charge_request.CreditCard.CreditCardExpirationMonth = 1;
+
+    strcpy(txn->charge_request.Amount.CurrencyCode, "USD");
+    txn->charge_request.Amount.Units = 300;
+    txn->charge_request.Amount.Nanos = 2;
 }
 
-static void SearchProducts(struct http_transaction *txn)
+static void PrintChargeResponse(struct http_transaction *txn)
 {
-    SearchProductsRequest *req = &txn->search_products_request;
-    SearchProductsResponse *out = &txn->search_products_response;
-    out->num_products = 0;
-
-    // Intepret query as a substring match in name or description.
-    int size = sizeof(product_ids) / sizeof(product_ids[0]);
-    int i = 0;
-    for (i = 0; i < size; i++)
-    {
-        if (strstr(products[i].Name, req->Query) != NULL || strstr(products[i].Description, req->Query) != NULL)
-        {
-            out->Results[out->num_products] = products[i];
-            out->num_products++;
-        }
-    }
-    return;
+    log_info("TransactionId: %s", txn->charge_response.TransactionId);
 }
 
 static void *nf_worker(void *arg)
@@ -224,34 +199,20 @@ static void *nf_worker(void *arg)
             return NULL;
         }
 
-        if (strcmp(txn->rpc_handler, "ListProducts") == 0)
+        if (strcmp(txn->rpc_handler, "Charge") == 0)
         {
-            ListProducts(txn);
-        }
-        else if (strcmp(txn->rpc_handler, "SearchProducts") == 0)
-        {
-            SearchProducts(txn);
-        }
-        else if (strcmp(txn->rpc_handler, "GetProduct") == 0)
-        {
-            GetProduct(txn);
+            Charge(txn);
         }
         else
         {
             log_info("%s() is not supported", txn->rpc_handler);
             log_info("\t\t#### Run Mock Test ####");
-            ListProducts(txn);
-            PrintListProductsResponse(txn);
-            MockGetProductRequest(txn);
-            GetProduct(txn);
-            PrintGetProductResponse(txn);
-            MockSearchProductsRequest(txn);
-            SearchProducts(txn);
-            PrintSearchProductsResponse(txn);
+            MockChargeRequest(txn);
+            PrintChargeResponse(txn);
         }
 
         txn->next_fn = txn->caller_fn;
-        txn->caller_fn = PRODUCTCATA_SVC;
+        txn->caller_fn = PAYMENT_SVC;
 
         bytes_written = write(pipefd_tx[index][1], &txn, sizeof(struct http_transaction *));
         if (unlikely(bytes_written == -1))
@@ -381,7 +342,7 @@ static int nf(uint8_t nf_id)
         return -1;
     }
 
-    cfg = memzone->addr;
+    cfg = (struct spright_cfg_s *)memzone->addr;
 
     ret = io_init();
     if (unlikely(ret == -1))
@@ -527,8 +488,6 @@ int main(int argc, char **argv)
         goto error_1;
     }
 
-    productcatalog_map = new_c_map(compare_e, NULL, NULL);
-    parseCatalog(productcatalog_map);
     ret = nf(nf_id);
     if (unlikely(ret == -1))
     {
